Add the background sound for WebGL
Implemented a separate version of the background sound specifically for the pause menu in WebGL. This was necessary because the Low Pass Filter used in the pause menu does not work in WebGL.
A new WebGLBgSound GameObject was created on Hierarchy to handle this sound.
The logic for changing the sound when paused and unpaused is now contained in MainUIHandler.cs.
I also corrected minor errors in the comments of the MainUIHandler.cs file.