Fix game quitting behavior in WebGL
I fixed a problem where the game wouldn't close properly when exiting in WebGL. Instead of closing, a blank page would open in the browser.

In the BaseUIHandler.cs script, I replaced the opening of the blank page with the opening of the main menu. This is a temporary fix until I find a better solution.